Adds test for should_notify_past_expiration
